rewrite uses of TodoError, moved setTimeout to ErroeNotification, moved user_id const
