+++ src/components/ErrorNotification/ErrorNotification.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React from 'react';
+import React, { useEffect } from 'react';
 import classNames from 'classnames';
 
 interface Props {
@@ -9,6 +9,13 @@ interface Props {
 
 export const ErrorNotification: React.FC<Props> = (props) => {
   const { errorMessage, setErrorHide } = props;
+
+  useEffect(() => {
+    const timerId = setTimeout(setErrorHide, 3000);
+
+    return () => clearTimeout(timerId);
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [errorMessage]);
 
   return (
     <div
