+++ src/utils/consts.ts
@@ -1,0 +1,1 @@
+export const USER_ID = 89;
